requirements.txt
python-telegram-bot==22.0
python-dotenv
aiohttp
google-generativeai
Pillow
httpx
openai
nest_asyncio
jsonrpyc
httpx-sse
